to do

+++ src/graphql/property/property.resolvers.js
@@ -2,27 +2,32 @@ import PropertyService from '../../services/PropertyService.js';
 
 const resolvers = {
   Query: {
-    properties: async (_, __, { user }) => {
-      if (!user) throw new Error('Unauthorized');
-      return await PropertyService.getAllProperties(user._id);
-    },
-    property: async (_, { id }, { user }) => {
-      if (!user) throw new Error('Unauthorized');
-      return await PropertyService.getPropertyById(id);
-    }
+    properties: async () => await PropertyService.getAllProperties(),
+
+    property: async (_, { id }) => await PropertyService.getPropertyById(id)
   },
+
   Mutation: {
-    createProperty: async (_, { title, description, price, category, location, status }, { user }) => {
+    createProperty: async (_, { type, title, description, bedrooms, bathrooms, size, orientation, block, price, location, project }, { user }) => {
       if (!user) throw new Error('Unauthorized');
-      return await PropertyService.createProperty({ title, description, price, category, location, status });
+      if (user.role === 'admin') {
+        const propertyData = { type, title, description, bedrooms, bathrooms, size, orientation, block, price, location, project };
+        return await PropertyService.createProperty(propertyData);
+      }
     },
+
     updateProperty: async (_, { id, ...args }, { user }) => {
       if (!user) throw new Error('Unauthorized');
-      return await PropertyService.updateProperty(id, args);
+      if (user.role === 'admin') {
+        return await PropertyService.updateProperty(id, args);
+      }
     },
+
     deleteProperty: async (_, { id }, { user }) => {
       if (!user) throw new Error('Unauthorized');
-      return await PropertyService.deleteProperty(id);
+      if (user.role === 'admin') {
+        return await PropertyService.deleteProperty(id);
+      }
     }
   }
 };
